refactor: header.js 분리
